+++ contracts/test/FnPtrsTest.sol
@@ -5,8 +5,8 @@ import {RainVM} from "../vm/RainVM.sol";
 import "../vm/VMStateBuilder.sol";
 
 uint256 constant ENTRYPOINT = 0;
-uint constant ENTRYPOINTS_LENGTH = 1;
-uint constant MIN_FINAL_STACK_INDEX = 1;
+uint256 constant ENTRYPOINTS_LENGTH = 1;
+uint256 constant MIN_FINAL_STACK_INDEX = 1;
 
 /// @title FnPtrsTest
 /// Test contract that returns bad fnPtrs length.
